requirements.txt
# Core dependencies
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

# AI and ML
sentence-transformers==2.2.2
pinecone-client==3.0.0

# Database (optional)
neo4j==5.9.0

# Utilities
tqdm==4.66.1
